Fixed issue with modifying WordLists
Fixed an issue where WordLists effectively stored two different lists (one as the instance variable self._collection, and one as the WordList itself, since it inherits the list class). This caused a variety of unexpected behavior, such as the pop method appearing not to modify the appearance of WordList, because it only modified the inherited list while the __str__ and __repr__ methods referenced the instance variable. Resolved by eliminating self._collections and modifying several methods to use the inherited list instead